make executable executable
Rework ram_storage, fix client waiting in server:wq

+++ inc/get_uptime.hpp
@@ -1,14 +1,13 @@
 #pragma once
 
 #include <cstdint>
-#include <array>
 #include <string>
 
 // using Name = std::array<char, 40>;
 using Name = std::string;
 struct ProcessInfo {
 	Name name{};
-	std::size_t uptime;
+	std::uint32_t uptime;
 };
 
 ProcessInfo FocusInfo();
